Split the attention test package
