+++ tests/test_flash_attention.py
@@ -1,0 +1,553 @@
+import math
+
+import numpy as np
+import pytest
+import torch
+import triton
+
+import flag_gems
+from flag_gems.runtime import torch_device_fn
+from flag_gems.utils import random_utils
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+device = flag_gems.device
+vendor_name = flag_gems.vendor_name
+
+
+def make_input(
+    batch,
+    num_head,
+    num_head_k,
+    q_seq_len,
+    kv_seq_len,
+    head_size,
+    dtype,
+    device,
+    requires_grad=False,
+):
+    random_utils.set_philox_state(1234567890, 0, device)
+    q_shape = (batch, num_head, q_seq_len, head_size)
+    kv_shape = (batch, num_head_k, kv_seq_len, head_size)
+    q = torch.empty(q_shape, dtype=dtype, device=device).uniform_(-0.05, 0.05)
+    k = torch.empty(kv_shape, dtype=dtype, device=device).uniform_(-0.05, 0.05)
+    v = torch.empty(kv_shape, dtype=dtype, device=device).uniform_(-0.05, 0.05)
+
+    if requires_grad:
+        q.requires_grad_()
+        k.requires_grad_()
+        v.requires_grad_()
+
+    return q, k, v
+
+
+def torch_flash_fwd(
+    q, k, v, scale, is_causal, dropout_p=0, return_debug_mask=False, **extra_kwargs
+):
+    q = q.transpose(1, 2)
+    k = k.transpose(1, 2)
+    v = v.transpose(1, 2)
+
+    (
+        out,
+        lse,
+        seed,
+        offset,
+        debug_softmax,
+    ) = torch.ops.aten._flash_attention_forward(
+        q,
+        k,
+        v,
+        None,
+        None,
+        q.shape[-3],
+        k.shape[-3],
+        dropout_p,
+        is_causal,
+        return_debug_mask,
+        scale=scale,
+        **extra_kwargs,
+    )
+
+    return out, lse, seed, offset, debug_softmax
+
+
+def gems_flash_fwd(
+    q, k, v, scale, is_causal, dropout_p=0, return_debug_mask=False, **extra_kwargs
+):
+    q = q.transpose(1, 2)
+    k = k.transpose(1, 2)
+    v = v.transpose(1, 2)
+    (
+        out,
+        lse,
+        seed,
+        offset,
+        debug_softmax,
+    ) = flag_gems.ops.flash_attention_forward(
+        q,
+        k,
+        v,
+        None,
+        None,
+        q.shape[-3],
+        k.shape[-3],
+        dropout_p,
+        is_causal,
+        return_debug_mask,
+        scale=scale,
+        **extra_kwargs,
+    )
+
+    return out, lse, seed, offset, debug_softmax
+
+
+def attn_bias_from_alibi_slopes(slopes, seqlen_q, seqlen_k, causal=False):
+    # batch, nheads = slopes.shape
+    device = slopes.device
+    slopes = slopes.unsqueeze(-1).unsqueeze(-1)
+    if causal:
+        return (
+            torch.arange(-seqlen_k + 1, 1, device=device, dtype=torch.float32) * slopes
+        )
+
+    row_idx = torch.arange(seqlen_q, device=device, dtype=torch.long).unsqueeze(-1)
+    col_idx = torch.arange(seqlen_k, device=device, dtype=torch.long)
+    relative_pos = torch.abs(row_idx + seqlen_k - seqlen_q - col_idx)
+    return -slopes * relative_pos.to(dtype=slopes.dtype)
+
+
+@pytest.mark.skipif(vendor_name == "metax", reason="TODOFIX")
+@pytest.mark.skipif(cfg.TO_CPU, reason="Unsupported in CPU mode")
+@pytest.mark.skipif(vendor_name == "hygon", reason="RuntimeError")
+@pytest.mark.skipif(vendor_name == "mthreads", reason="Unsupported in CPU mode")
+@pytest.mark.flash_attention_forward
+@pytest.mark.parametrize(
+    ["batch", "num_head", "q_seq_len", "kv_seq_len"],
+    [(1, 1, 128, 2048), (4, 8, 1024, 128), (4, 8, 17, 1030)],
+)
+@pytest.mark.parametrize("head_size", [64, 128, 192, 256])
+@pytest.mark.parametrize("is_causal", [False, True])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
+def test_flash_attention_foward_nonsquare_qk(
+    batch, num_head, q_seq_len, kv_seq_len, head_size, is_causal, dtype
+):
+    device = torch_device_fn.current_device()
+    q, k, v = make_input(
+        batch, num_head, num_head, q_seq_len, kv_seq_len, head_size, dtype, device
+    )
+    ref_q = utils.to_reference(q, False)
+    ref_k = utils.to_reference(k, False)
+    ref_v = utils.to_reference(v, False)
+    scale = float(1.0 / np.sqrt(head_size))
+
+    torch_out, torch_lse, _, _, _ = torch_flash_fwd(
+        ref_q, ref_k, ref_v, scale, is_causal
+    )
+    gems_out, gems_lse, _, _, _ = gems_flash_fwd(q, k, v, scale, is_causal)
+
+    utils.gems_assert_close(gems_out, torch_out, dtype)
+    # TODO(Iluvatar): Don't return prematurily here.
+    if vendor_name == "iluvatar":
+        return
+    utils.gems_assert_close(gems_lse, torch_lse, torch.float)
+
+
+# Adapted from https://github.com/Dao-AILab/flash-attention/blob/main/tests/test_flash_attn.py
+def construct_local_mask(
+    seqlen_q,
+    seqlen_k,
+    window_size=(-1, -1),  # -1 means infinite window size
+    query_padding_mask=None,
+    key_padding_mask=None,
+    device=None,
+    key_leftpad=None,
+):
+    # row_idx = rearrange(torch.arange(seqlen_q, device=device, dtype=torch.long), "s -> s 1")
+    row_idx = torch.arange(seqlen_q, device=device, dtype=torch.long)[:, None]
+    col_idx = torch.arange(seqlen_k, device=device, dtype=torch.long)
+    if key_leftpad is not None:
+        # key_leftpad = rearrange(key_leftpad, "b -> b 1 1 1")
+        key_leftpad = key_leftpad[:, None, None, None]
+        # col_idx = repeat(col_idx, "s -> b 1 1 s", b=key_leftpad.shape[0])
+        col_idx = col_idx.repeat(key_leftpad.shape[0], 1, 1, 1)
+        col_idx = torch.where(col_idx >= key_leftpad, col_idx - key_leftpad, 2**32)
+    sk = (
+        seqlen_k
+        if key_padding_mask is None
+        # else rearrange(key_padding_mask.sum(-1), "b -> b 1 1 1")
+        else key_padding_mask.sum(-1)[:, None, None, None]
+    )
+    sq = (
+        seqlen_q
+        if query_padding_mask is None
+        # else rearrange(query_padding_mask.sum(-1), "b -> b 1 1 1")
+        else query_padding_mask.sum(-1)[:, None, None, None]
+    )
+    if window_size[0] < 0:
+        return col_idx > row_idx + sk - sq + window_size[1]
+    else:
+        sk = torch.full_like(col_idx, seqlen_k) if key_padding_mask is None else sk
+        return torch.logical_or(
+            col_idx > torch.minimum(row_idx + sk - sq + window_size[1], sk),
+            col_idx < row_idx + sk - sq - window_size[0],
+        )
+
+
+def attention_ref(
+    q,
+    k,
+    v,
+    scale,
+    query_padding_mask=None,
+    key_padding_mask=None,
+    attn_bias=None,
+    dropout_p=0.0,
+    dropout_mask=None,
+    causal=False,
+    window_size=(-1, -1),  # -1 means infinite window size
+    softcap=0.0,
+    upcast=True,
+    reorder_ops=False,
+    key_leftpad=None,
+):
+    """
+    Arguments:
+        q: (batch_size, seqlen_q, nheads, head_dim)
+        k: (batch_size, seqlen_k, nheads_k, head_dim)
+        v: (batch_size, seqlen_k, nheads_k, head_dim)
+        scale: float
+        query_padding_mask: (batch_size, seqlen_q)
+        key_padding_mask: (batch_size, seqlen_k)
+        attn_bias: broadcastable to (batch_size, nheads, seqlen_q, seqlen_k)
+        dropout_p: float
+        dropout_mask: (batch_size, nheads, seqlen_q, seqlen_k)
+        causal: whether to apply causal masking
+        window_size: (int, int), left and right window size
+        upcast: whether to cast all inputs to fp32, do all computation in fp32, then cast
+            output back to fp16/bf16.
+        reorder_ops: whether to change the order of operations (scaling k instead of scaling q, etc.)
+            without changing the math. This is to estimate the numerical error from operation
+            reordering.
+    Output:
+        output: (batch_size, seqlen_q, nheads, head_dim)
+        attention: (batch_size, nheads, seqlen_q, seqlen_k), softmax after dropout
+    """
+
+    q = q.transpose(1, 2)
+    k = k.transpose(1, 2)
+    v = v.transpose(1, 2)
+
+    q *= scale
+
+    if causal:
+        window_size = (window_size[0], 0)
+    dtype_og = q.dtype
+    if upcast:
+        q, k, v = q.float(), k.float(), v.float()
+    seqlen_q, seqlen_k = q.shape[1], k.shape[1]
+    g = q.shape[2] // k.shape[2]
+    # k = repeat(k, "b s h d -> b s (h g) d", g=q.shape[2] // k.shape[2])
+    # v = repeat(v, "b s h d -> b s (h g) d", g=q.shape[2] // v.shape[2])
+    k = k.repeat_interleave(g, dim=2)
+    v = v.repeat_interleave(g, dim=2)
+    d = q.shape[-1]
+    if not reorder_ops:
+        scores = torch.einsum("bthd,bshd->bhts", q / math.sqrt(d), k)
+    else:
+        scores = torch.einsum("bthd,bshd->bhts", q, k / math.sqrt(d))
+
+    if softcap > 0:
+        scores = scores / softcap
+        scores = scores.tanh()
+        scores = scores * softcap
+
+    if key_padding_mask is not None:
+        scores.masked_fill_((~key_padding_mask)[:, None, None, :], float("-inf"))
+
+    if window_size[0] >= 0 or window_size[1] >= 0:
+        local_mask = construct_local_mask(
+            seqlen_q,
+            seqlen_k,
+            window_size,
+            query_padding_mask,
+            key_padding_mask,
+            q.device,
+            key_leftpad=key_leftpad,
+        )
+        scores.masked_fill_(local_mask, float("-inf"))
+    if attn_bias is not None:
+        scores = scores + attn_bias
+    attention = torch.softmax(scores, dim=-1).to(v.dtype)
+    # Some rows might be completely masked out so we fill them with zero instead of NaN
+    if window_size[0] >= 0 or window_size[1] >= 0:
+        attention = attention.masked_fill(
+            torch.all(local_mask, dim=-1, keepdim=True), 0.0
+        )
+    # We want to mask here so that the attention matrix doesn't have any NaNs
+    # Otherwise we'll get NaN in dV
+    if query_padding_mask is not None:
+        mask = (~query_padding_mask)[:, None, :, None]
+        attention = attention.masked_fill(mask, 0.0)
+
+    dropout_scaling = 1.0 / (1 - dropout_p)
+    # attention_drop = attention.masked_fill(~dropout_mask, 0.0) * dropout_scaling
+    # output = torch.einsum('bhts,bshd->bthd', attention_drop , v)
+    if dropout_mask is not None:
+        attention_drop = attention.masked_fill(~dropout_mask, 0.0)
+    else:
+        attention_drop = attention
+    output = torch.einsum("bhts,bshd->bthd", attention_drop, v * dropout_scaling)
+    if query_padding_mask is not None:
+        output.masked_fill_((~query_padding_mask)[:, :, None, None], 0.0)
+    return output.to(dtype=dtype_og), attention.to(dtype=dtype_og)
+
+
+@pytest.mark.skipif(cfg.TO_CPU, reason="Unsupported in CPU mode")
+@pytest.mark.skipif(vendor_name == "hygon", reason="RuntimeError")
+@pytest.mark.skipif(vendor_name == "mthreads", reason="Unsupported in CPU mode")
+@pytest.mark.skipif(vendor_name == "kunlunxin", reason="RESULT TODOFIX")
+@pytest.mark.flash_attention_forward
+@pytest.mark.parametrize(
+    ["batch", "num_head", "num_head_k", "q_seq_len", "kv_seq_len"],
+    [(4, 8, 2, 1024, 1024), (4, 4, 4, 1, 519)],
+)
+@pytest.mark.parametrize("head_size", [128, 192])
+@pytest.mark.parametrize("is_causal", [False, True])
+@pytest.mark.parametrize("soft_cap", [None, 10.0, 50.0])
+@pytest.mark.parametrize("alibi", [True])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
+def test_flash_attention_forward_gqa_alibi_softcap(
+    batch,
+    num_head,
+    num_head_k,
+    q_seq_len,
+    kv_seq_len,
+    head_size,
+    is_causal,
+    soft_cap,
+    alibi,
+    dtype,
+):
+    device = torch_device_fn.current_device()
+    q, k, v = make_input(
+        batch, num_head, num_head_k, q_seq_len, kv_seq_len, head_size, dtype, device
+    )
+    ref_q = utils.to_reference(q, False)
+    ref_k = utils.to_reference(k, False)
+    ref_v = utils.to_reference(v, False)
+    scale = float(1.0 / np.sqrt(head_size))
+
+    if alibi:
+        # alibi_slopes = torch.rand(batch, num_head, device=device, dtype=torch.float32) * 0.3
+        alibi_slopes = (
+            torch.ones(batch, num_head, device=device, dtype=torch.float32) * 0.3
+        )
+        attn_bias = attn_bias_from_alibi_slopes(
+            alibi_slopes, q_seq_len, kv_seq_len, causal=is_causal
+        )
+    else:
+        alibi_slopes, attn_bias = None, None
+
+    torch_out, _ = attention_ref(
+        ref_q,
+        ref_k,
+        ref_v,
+        scale,
+        None,
+        None,
+        attn_bias,
+        0.0,
+        None,
+        causal=is_causal,
+        window_size=(-1, -1),
+        softcap=soft_cap if soft_cap is not None else 0,
+    )
+
+    gems_out, _, _, _, _ = gems_flash_fwd(
+        q,
+        k,
+        v,
+        scale,
+        is_causal,
+        alibi_slopes=alibi_slopes,
+        softcap=soft_cap if soft_cap is not None else 0,
+        disable_splitkv=True,
+    )
+
+    utils.gems_assert_close(gems_out, torch_out, dtype)
+
+
+@pytest.mark.skipif(vendor_name == "metax", reason="TODOFIX")
+@pytest.mark.skipif(cfg.TO_CPU, reason="Unsupported in CPU mode")
+@pytest.mark.skipif(vendor_name == "hygon", reason="RuntimeError")
+@pytest.mark.skipif(vendor_name == "mthreads", reason="Unsupported in CPU mode")
+@pytest.mark.skipif(vendor_name == "kunlunxin", reason="RESULT TODOFIX")
+@pytest.mark.flash_attention_forward
+@pytest.mark.parametrize(
+    ["batch", "num_head", "num_head_k", "q_seq_len", "kv_seq_len"],
+    [(1, 4, 1, 1, 1024), (4, 4, 4, 1, 519)],
+)
+@pytest.mark.parametrize("head_size", [128, 192])
+@pytest.mark.parametrize("is_causal", [False, True])
+@pytest.mark.parametrize("soft_cap", [None, 10.0, 50.0])
+@pytest.mark.parametrize("alibi", [False, True])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
+def test_flash_attention_foward_splitkv(
+    batch,
+    num_head,
+    num_head_k,
+    q_seq_len,
+    kv_seq_len,
+    head_size,
+    is_causal,
+    soft_cap,
+    alibi,
+    dtype,
+):
+    device = torch_device_fn.current_device()
+    q, k, v = make_input(
+        batch, num_head, num_head_k, q_seq_len, kv_seq_len, head_size, dtype, device
+    )
+    ref_q = utils.to_reference(q, False)
+    ref_k = utils.to_reference(k, False)
+    ref_v = utils.to_reference(v, False)
+    scale = float(1.0 / np.sqrt(head_size))
+
+    if alibi:
+        # alibi_slopes = torch.rand(batch, num_head, device=device, dtype=torch.float32) * 0.3
+        alibi_slopes = (
+            torch.ones(batch, num_head, device=device, dtype=torch.float32) * 0.3
+        )
+        attn_bias = attn_bias_from_alibi_slopes(
+            alibi_slopes, q_seq_len, kv_seq_len, causal=is_causal
+        )
+    else:
+        alibi_slopes, attn_bias = None, None
+
+    torch_out, _ = attention_ref(
+        ref_q,
+        ref_k,
+        ref_v,
+        scale,
+        None,
+        None,
+        attn_bias,
+        0.0,
+        None,
+        causal=is_causal,
+        window_size=(-1, -1),
+        softcap=soft_cap if soft_cap is not None else 0,
+    )
+
+    gems_out, gems_lse, _, _, _ = gems_flash_fwd(
+        q,
+        k,
+        v,
+        scale,
+        is_causal,
+        alibi_slopes=alibi_slopes,
+        softcap=soft_cap if soft_cap is not None else 0,
+    )
+
+    utils.gems_assert_close(gems_out, torch_out, dtype)
+
+
+@pytest.mark.skipif(vendor_name == "metax", reason="TODOFIX")
+@pytest.mark.skipif(cfg.TO_CPU, reason="Unsupported in CPU mode")
+@pytest.mark.skipif(torch.__version__ < "2.4", reason="Low Pytorch Version")
+@pytest.mark.skipif(vendor_name == "hygon", reason="RuntimeError")
+@pytest.mark.skipif(vendor_name == "mthreads", reason="Unsupported in CPU mode")
+@pytest.mark.skipif(vendor_name == "kunlunxin", reason="RESULT TODOFIX")
+@pytest.mark.flash_attention_forward
+@pytest.mark.parametrize(
+    ["batch", "num_head", "q_seq_len", "kv_seq_len"],
+    [(1, 1, 128, 2048), (8, 32, 1024, 1024), (8, 32, 1024, 128), (8, 32, 17, 1030)],
+)
+@pytest.mark.parametrize("head_size", [128, 192])
+@pytest.mark.parametrize(
+    ["window_size_left", "window_size_right"], [(256, 0), (128, 128)]
+)
+@pytest.mark.parametrize("is_causal", [False])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
+def test_flash_attention_foward_swa(
+    batch,
+    num_head,
+    q_seq_len,
+    kv_seq_len,
+    head_size,
+    is_causal,
+    window_size_left,
+    window_size_right,
+    dtype,
+):
+    device = torch_device_fn.current_device()
+    q, k, v = make_input(
+        batch, num_head, num_head, q_seq_len, kv_seq_len, head_size, dtype, device
+    )
+    ref_q = utils.to_reference(q, False)
+    ref_k = utils.to_reference(k, False)
+    ref_v = utils.to_reference(v, False)
+    scale = float(1.0 / np.sqrt(head_size))
+
+    torch_out, torch_lse, _, _, _ = torch_flash_fwd(
+        ref_q,
+        ref_k,
+        ref_v,
+        scale,
+        is_causal,
+        dropout_p=0,
+        return_debug_mask=False,
+        window_size_left=window_size_left,
+        window_size_right=window_size_right,
+    )
+    gems_out, gems_lse, _, _, _ = gems_flash_fwd(
+        q,
+        k,
+        v,
+        scale,
+        is_causal,
+        dropout_p=0,
+        return_debug_mask=False,
+        window_size_left=window_size_left,
+        window_size_right=window_size_right,
+    )
+
+    utils.gems_assert_close(gems_out, torch_out, dtype)
+    # TODO(Iluvatar): Don't return early here.
+    if vendor_name == "iluvatar":
+        return
+    utils.gems_assert_close(gems_lse, torch_lse, torch.float)
+
+
+@pytest.mark.skipif(cfg.TO_CPU, reason="Unsupported in CPU mode")
+@pytest.mark.skipif(triton.__version__ < "3.1", reason="Low Triton Version")
+@pytest.mark.skipif(vendor_name == "hygon", reason="RuntimeError")
+@pytest.mark.skipif(vendor_name == "mthreads", reason="Low Triton Version")
+@pytest.mark.skipif(vendor_name == "kunlunxin", reason="RESULT TODOFIX")
+@pytest.mark.flash_attention_forward
+@pytest.mark.parametrize(
+    ["batch", "num_head", "q_seq_len", "kv_seq_len"],
+    [
+        (1, 1, 1024, 1024),
+    ],
+)
+@pytest.mark.parametrize("head_size", [128])
+@pytest.mark.parametrize("is_causal", [False, True])
+@pytest.mark.parametrize("dtype", [torch.float16, torch.bfloat16])
+def test_flash_fwd_dropout(
+    batch, num_head, q_seq_len, kv_seq_len, head_size, is_causal, dtype
+):
+    device = torch_device_fn.current_device()
+    q, k, v = make_input(
+        batch, num_head, num_head, q_seq_len, kv_seq_len, head_size, dtype, device
+    )
+    scale = float(1.0 / np.sqrt(head_size))
+    dropout_p = 0.2
+    _, _, _, _, debug_softmax = gems_flash_fwd(
+        q, k, v, scale, is_causal, dropout_p=dropout_p, return_debug_mask=True
+    )
+
+    dropout_ratio = torch.sum(debug_softmax < 0) / torch.sum(debug_softmax != 0)
+    np.testing.assert_allclose(dropout_ratio.to("cpu"), dropout_p, rtol=5e-2)
